+++ code/parachain/runtime/dali/src/prelude.rs
@@ -1,0 +1,3 @@
+pub use sp_std::prelude::*;
+
+pub use frame_support::{traits::PalletInfoAccess, weights::Weight};
